Today I have done animations with the screen and completed icons for veg-non veg and implementation of TextView below Image View forminimum order of resturant and current continuing with Search Location functionality by refering google references.
22-5-2017
Sir made design for search bar
Completed Updation of amount of item
Alert Dialog if the quantity goes higher than 10 with validation

23-5-2017
Completed Search functionality for location

24-5-2017
Creating bill for resturant

25-5-2017
Designing Screen for bill payment
and create layout resource file to inflate it

26-5-2017
Created bill by inflating layout in listview and bill will be updated soon

27-5-2017
Creation of bill completed with help of pratik sir
Layout designing will be completed by tommorrow

28-5-2017
Designing layout for bill

29-5-2017
Continuing with the designing of bill

31-5-2017
Added Names of Resturants in ExpandableListView
Created Icon For Veg/Non Veg
and also added icons dynamically according to the types of resturants

1-6-2017
Try to get details of restuarant restuarant wise but not getting it
now starting to design screen for filter
and icon of filter added on the action bar

2-6-2017
Data getting appropriately according to Resturants not getting amount and item count
continuing screen designing for filters in resturants criteria

3-6-2017
Completed Resturant wise order validation
and continuing with implementation of FilterScreen using RecylerView

5-6-2017
Completed Screen for Filter and also implementatio of On/Off button
for filter according to Rating and Delivery

6-6-2017
Completed  code to find latitude and longtitude according to place selected by the user

7-6-2017
Added Search Bar to search restaurants(only design)
Validation According to minimum order value completed
Added data from adapter of categories with appropriate checkbox

8-6-2017
Completed Search Fucnctionality for the Restaurant and designing of bill

9-6-2017
Completed UI for cart,Added item counts and Sub price of Restaurants
and also validate amount for Minimum value and restaurant order(Red,Green)

10-6-2017
Completed coding for updating the location according to select on user

12-6-2017
Tried for creating Collapse Layout of Toolbar for adding the image of the Restaurant but its not happening because of Relative Layout taken on the screen of Menu
and we have to Compulsory take Coordinate Layout which is not possible because of BottomNavigation View

13-6-2017
Trying to get Restaurants timing from json but not getting
added  restuarant name and its timings textView in ToolBar

14-6-2017
Getting Timings of Service according to restaurants from web service
validations will be continued tommorrow....

15-6-2017
Added Timings on the restaurannt screen

16-6-2017
Perfectly added timings to the list of restaurrants according to approach of Fajar
and getting current time of the system....

21-6-2017
Completed Status of Restaurant validation with its timings open or close in android studio

+++ app/src/main/java/com/goyo/grocery_goyo/model/restaurantModel.java
@@ -51,7 +51,7 @@ public class restaurantModel {
     public String resttype;
 
    @SerializedName("minordamt")
-    public Double min_order;
+    public Integer min_order;
     public Integer getRestid() {
         return restid;
     }
@@ -124,10 +124,10 @@ public class restaurantModel {
         this.resttype = resttype;
     }
 
-    public Double getMin_order() {
+    public Integer getMin_order() {
         return min_order;
     }
-    public void setMin_order(Double min_order) {
+    public void setMin_order(Integer min_order) {
         this.min_order = min_order;
     }
